:truck: Rename js files to jsx
