+++ src/pages/detect/detect.jsx
