+++ src/pages/train/train.jsx
